(Jade) Some corrections on multicore

+++ projects/Jade/src/orcc/orcc/src/display.c
@@ -80,7 +80,7 @@ static void displayYUV_setSize(int width, int height) {
 
 void displayYUV_displayPicture(unsigned char *pictureBufferY,
 		unsigned char *pictureBufferU, unsigned char *pictureBufferV,
-		unsigned int   pictureWidth, unsigned int pictureHeight) {
+		unsigned short   pictureWidth, unsigned short pictureHeight) {
 	static unsigned int lastWidth = 0;
 	static unsigned int lastHeight = 0;
 	SDL_Event event;
